Move inside of component folder
