+++ src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { BookDetailsComponent } from '../app/components/book-details/book-details.component';
 import { BookListComponent } from '../app/components/book-list/book-list.component';
-import { CheckoutComponent } from './checkout/checkout.component';
+import { CheckoutComponent } from '../app/components/checkout/checkout.component';
 import { CartComponent } from './components/cart/cart.component';
 
 const routes: Routes = [
